+++ src/lib.rs
@@ -127,7 +127,7 @@ pub use chromatic_adaptation_transform::{Cat, ChromaticAdaptationTransform};
 pub use context::ColorimetricContext;
 pub use error::Error;
 pub use illuminant::{Illuminant, IlluminantType};
-pub use observer::Observer;
+pub use observer::{Builder as ObserverBuilder, Modifier as FairchildModifier, Observer};
 pub use spectral::{
   ChromaticityCoordinates, Cmf, ColorMatchingFunction, ConeFundamentals, ConeResponse, Spd, SpectralPowerDistribution,
   Table as SpectralTable, TristimulusResponse,